Refactoring e modifiche minori applicazione articoli
