+++ articoli/templates/dettaglio_articolo.html
@@ -1,0 +1,4 @@
+<p>{{articolo.titolo}}</p>
+<p>{{articolo.data_inizio_pubblicazione}}</p>
+<p>{{articolo.corpo|safe}}</p>
+<p>{{articolo.visualizzazioni}}</p>
